deletion confirmation

+++ source/js/project.js
@@ -66,13 +66,17 @@ function createProject() {
  * @param {string} projectId string identifier of the project to be deleted
  */
 function deleteProject(projectId) {
-  let parentId = document.getElementById(projectId);
-  projects = projects.filter((project) => project != projectId); // removes project from projects array
+  if (window.confirm(`Are you sure you want to delete ${projectId}? (This action cannot be undone)`)) {
+    let parentId = document.getElementById(projectId);
+    projects = projects.filter((project) => project != projectId); // removes project from projects array
 
-  const projectList = document.getElementById("Project-List");
-  projectList.removeChild(parentId); // removing project visually from website
+    const projectList = document.getElementById("Project-List");
+    projectList.removeChild(parentId); // removing project visually from website
 
-  localStorage.setItem("projects", JSON.stringify(projects)); // saves the projects in local storage
+    localStorage.setItem("projects", JSON.stringify(projects)); // saves the projects in local storage
+  } else {
+    return;
+  }
 }
 
 /**
